changed app.ts and index.ts file logic to enable supertests
